Fix build on FreeBSD and compatability change
Use int as it should be in mk-configure prototype for FreeBSD.
Fix typo in cmake protype lookup for FreeBSD.
Comptability change, Use off_t as type for the amount of written bytes.
This is safe since we are reading from files and off_t is the type for file
offsets.
Include netinet/in.h to make sure sockaddr_in if defined.